fix (InputManager) : Commentator 클래스 기능 분리
