+++ src/main/java/racingcar/controller/RaceController.java
@@ -5,24 +5,27 @@ import java.util.List;
 import racingcar.domain.Car;
 import racingcar.service.RaceManager;
 import racingcar.view.Commentator;
+import racingcar.view.InputManager;
 
 public class RaceController {
 	private final RaceManager raceManager;
 	private final Commentator commentator;
+	private final InputManager inputManager;
 
 	public RaceController() {
 		raceManager = new RaceManager();
 		commentator = new Commentator();
+		inputManager = new InputManager();
 	}
 
 	public void start() {
-		List<Car> cars = commentator.getCarNames();
-		int number = commentator.getRuntimeNumber();
+		List<Car> cars = inputManager.getCars();
+		int number = inputManager.getRuntimeNumber();
 		startRace(cars, number);
 	}
 
 	private void startRace(List<Car> cars, int runtimeNumber) {
-		commentator.printRaceStart();
+		commentator.notifyRaceResult();
 		for(int i = 0; i < runtimeNumber; i++) {
 			raceManager.moveCars(cars);
 			commentator.notifyCarsStatement(cars);
