+++ src/main/java/racingcar/service/Converter.java
@@ -4,12 +4,12 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import racingcar.constants.Constants;
+import racingcar.constants.RaceConstants;
 import racingcar.domain.Car;
 
 public class Converter {
 	public List<String> convertStringToList(String inputString) {
-		String [] splitedArray = inputString.split(Constants.CUT_OFF_POINT);
+		String [] splitedArray = inputString.split(RaceConstants.CUT_OFF_POINT);
 		return Arrays.asList(splitedArray);
 	}
